+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Support translate and transform quantization metadata
+
 ## [2.7.0] - 11-20-2017
 ### Added
 * `filtersApplied.all skips all post processing`